feat(layout-#8-#9) Implementation of the navigation menu, header, footer in the main Twig template
