+++ templates/base.html.twig
@@ -4,7 +4,7 @@
         <meta charset="UTF-8" />
         <meta http-esquiv="X-UA-Compatible" content="IE-Edge">
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>{% block title %}Blog{% endblock %}</title>
+        <title>{% block title %}Blog de John Doe{% endblock %}</title>
         <!-- Favicon-->
         <link rel="icon" type="image/x-icon" href="../assets/favicon.ico" />
         <!-- Font Awesome icons -->
@@ -12,16 +12,88 @@
         <!-- Google fonts-->
         <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700" rel="stylesheet" type="text/css" />
         <link href="https://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic" rel="stylesheet" type="text/css" />
-        <!-- CSS -->
-        {% block stylesheets %}
-            <link href="../public/assets/css/styles.css" rel="stylesheet" />
-        {% endblock %}
+        <!-- Bootstrap CSS -->
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
+        <!-- Custom CSS -->
+        <link href="assets/css/styles.css" rel="stylesheet" />
     </head>
-    <body>        
-        {% block body %}{% endblock %}
-        <!-- Bootstrap core JS-->
+    <body>   
+        <!-- Navigation-->
+        <nav class="nav navbar navbar-expand-lg text-uppercase fixed-top" id="mainNav">
+            <div class="container">
+                <a class="navbar-brand" href="#">Blog de John Doe</a>
+                <button class="navbar-toggler text-uppercase font-weight-bold text-white rounded" type="button" data-bs-toggle="collapse" data-bs-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
+                    Menu
+                    <i class="fas fa-bars"></i>
+                </button>
+                <div class="collapse navbar-collapse" id="navbarResponsive">
+                    <ul class="navbar-nav ms-auto">
+                        <li class="nav-item mx-0 mx-lg-1"><a class="nav-link py-3 px-0 px-lg-3 rounded" href="#">Accueil</a></li>
+                        <li class="nav-item mx-0 mx-lg-1"><a class="nav-link py-3 px-0 px-lg-3 rounded" href="#">Articles</a></li>
+                        <li class="nav-item mx-0 mx-lg-1"><a class="nav-link py-3 px-0 px-lg-3 rounded" href="#">Connexion</a></li>
+                        <li class="nav-item mx-0 mx-lg-1"><a class="nav-link py-3 px-0 px-lg-3 rounded" href="#">Inscription</a></li>
+                    </ul>
+                </div>
+            </div>
+        </nav>
+
+        <!-- Header-->
+        <header class="masthead text-white text-center">
+            <div class="container d-flex align-items-center flex-column">
+                <img class="masthead-avatar mb-5" src="../assets/img/avataaars.svg" alt="..." />
+                <h1 class="masthead-heading text-uppercase mb-0">John Doe</h1>
+                <div class="divider-custom divider-light">
+                    <div class="divider-custom-line"></div>
+                    <div class="divider-custom-icon"><i class="fas fa-star"></i></div>
+                    <div class="divider-custom-line"></div>
+                </div>
+                <p class="masthead-subheading font-weight-light mb-0">Le développeur dont vous avez besoin !</p>
+            </div>
+        </header>
+
+        <!-- Scroll to top -->
+        <button onclick="topFunction()" id="scrollToTop"><i class="fa-solid fa-chevron-up"></i></button> 
+
+        <!-- Main -->
+        {% block main %}{% endblock %}
+
+        <!-- Footer-->
+        <footer class="footer text-center">
+            <div class="container">
+                <div class="row">
+                    <div class="col-lg-4 mb-5 mb-lg-0">
+                        <h4 class="text-uppercase mb-4">Localisation</h4>
+                        <p class="footer-localisation lead mb-0">
+                            250 Rue de Paris
+                            <br />
+                            Paris, 75000
+                        </p>
+                    </div>
+                    <div class="col-lg-4 mb-5 mb-lg-0">
+                        <h4 class="text-uppercase mb-4">Suivez-moi sur les réseaux</h4>
+                        <a class="btn btn-social mx-1" href="https://www.linkedin.com/in/pgael/" target="_blank"><i class="footer-icon fab fa-fw fa-linkedin-in"></i></a>
+                        <a class="btn btn-social mx-1" href="https://github.com/Galusss" target="_blank"><i class="footer-icon fab fa-fw fa-github"></i></a>
+                        <a class="btn btn-social mx-1" href="https://twitter.com/galuss_" target="_blank"><i class="footer-icon fab fa-fw fa-twitter"></i></a>
+                    </div>
+                    <div class="col-lg-4">
+                        <h4 class="text-uppercase mb-4">À propos</h4>
+                        <ul class="footer-list lead mb-0">
+                            <li><a href="#">Conditions générales d'utilisation</a></li>
+                            <li><a href="#">Politique de confidentialité</a></li>
+                        </ul>
+                    </div>
+                </div>
+            </div>
+        </footer>
+
+        <!-- Copyright -->
+        <div class="copyright py-4 text-center text-white">
+            <div class="container"><span>Copyright 2022 &copy; Blog de John Doe</span></div>
+        </div>
+
+        <!-- Bootstrap JS -->
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-        <!-- Core theme JS-->
-        <script src="../assets/js/index.js"></script>
+        <!-- Custom JS -->
+        <script src="assets/js/index.js"></script>
     </body>
 </html>
