+++ templates/posts/index.html.twig
@@ -1,5 +1,12 @@
 {% extends "base.html.twig" %}
 
-{% block body %}
-    <h1>Hello World</h1>
+{% block main %}
+    <h1 class="colorA">Tan</h1>
+    <br>
+    <h1 class="colorB">Cream</h1>
+    <br>
+    <h1 class="colorC">Charocal</h1>
+    <br>
+    <h1 class="colorD">Gold</h1>
+    <br>
 {% endblock %}
